LayoutTests/svg/dom/valueAsString-null-expected.txt

Test assigning null to SVGAngle.valueAsString / SVGLength.valueAsString

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS angle.valueAsString is "0"
PASS angle.valueAsString = null threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS angle.valueAsString is "0"
PASS angle.valueAsString = '10' did not throw exception.
PASS angle.valueAsString is "10"
FAIL angle.valueAsString = '' should throw an exception. Was .
PASS angle.valueAsString is "10"

PASS length.valueAsString is "0"
PASS length.valueAsString = null threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS length.valueAsString is "0"
PASS length.valueAsString = '10' did not throw exception.
PASS length.valueAsString is "10"
FAIL length.valueAsString = '' should throw an exception. Was .
PASS length.valueAsString is "10"
PASS successfullyParsed is true

TEST COMPLETE
